+++ apm-collector/apm-collector-agentstream/src/test/java/org/skywalking/apm/collector/agentstream/mock/SegmentPost.java
@@ -135,8 +135,4 @@ public class SegmentPost {
 
         return osInfoJson;
     }
-
-    private void newDao() {
-
-    }
 }
