+++ apm-collector/apm-collector-cache/src/main/java/org/skywalking/apm/collector/cache/dao/InstanceEsCacheDAO.java
@@ -18,11 +18,25 @@
 
 package org.skywalking.apm.collector.cache.dao;
 
+import org.elasticsearch.action.get.GetResponse;
+import org.skywalking.apm.collector.storage.define.register.InstanceTable;
+import org.skywalking.apm.collector.storage.elasticsearch.dao.EsDAO;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
 /**
  * @author pengys5
  */
-public interface IApplicationDAO {
-    int getApplicationId(String applicationCode);
+public class InstanceEsCacheDAO extends EsDAO implements IInstanceCacheDAO {
 
-    String getApplicationCode(int applicationId);
+    private final Logger logger = LoggerFactory.getLogger(InstanceEsCacheDAO.class);
+
+    @Override public int getApplicationId(int applicationInstanceId) {
+        GetResponse response = getClient().prepareGet(InstanceTable.TABLE, String.valueOf(applicationInstanceId)).get();
+        if (response.isExists()) {
+            return (int)response.getSource().get(InstanceTable.COLUMN_APPLICATION_ID);
+        } else {
+            return 0;
+        }
+    }
 }
