+++ apm-collector/apm-collector-storage/src/main/java/org/skywalking/apm/collector/storage/h2/dao/H2DAO.java
@@ -20,7 +20,6 @@ package org.skywalking.apm.collector.storage.h2.dao;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
-
 import org.skywalking.apm.collector.client.h2.H2Client;
 import org.skywalking.apm.collector.client.h2.H2ClientException;
 import org.skywalking.apm.collector.storage.dao.DAO;
@@ -32,17 +31,18 @@ import org.slf4j.LoggerFactory;
  */
 public abstract class H2DAO extends DAO<H2Client> {
     private final Logger logger = LoggerFactory.getLogger(H2DAO.class);
-    public final int getMaxId(String tableName, String columnName) {
+
+    protected final int getMaxId(String tableName, String columnName) {
         String sql = "select max(" + columnName + ") from " + tableName;
         return getIntValueBySQL(sql);
     }
 
-    public final int getMinId(String tableName, String columnName) {
+    protected final int getMinId(String tableName, String columnName) {
         String sql = "select min(" + columnName + ") from " + tableName;
         return getIntValueBySQL(sql);
     }
 
-    public final int getIntValueBySQL(String sql) {
+    private int getIntValueBySQL(String sql) {
         H2Client client = getClient();
         try (ResultSet rs = client.executeQuery(sql, null)) {
             if (rs.next()) {
@@ -58,5 +58,4 @@ public abstract class H2DAO extends DAO<H2Client> {
         }
         return 0;
     }
-
 }
